Change "Duration is not aligned properly" message from Warning to Info
This message appears whenever P4G starts if any mod has an audio track that is not exactly the same duration as the track it is replacing.
It is confusing users who believe something is wrong with their modding setup and has no value to anyone except creators, who should be testing their mods before release regardless.